+++ src/features/battery/data.rs
@@ -8,7 +8,7 @@ use utils::icon_by_percentage;
 
 #[derive(Clone, Debug)]
 pub(super) struct BatteryInfo {
-    pub(super) capacity: u16,
+    pub(super) capacity: u32,
     pub(super) estimation: Option<time::Duration>,
 }
